Eviter crash server pour un bug suppression supplémentaire

+++ server/app/services/database.games.service.ts
@@ -38,7 +38,7 @@ export class RecordTimesService {
     }
 
     async deleteGameRecordTimes(nameOfWantedGame: string): Promise<void> {
-        if (this.isDatabaseAvailable()) {
+        if (this.isDatabaseAvailable() && !(await this.validateName(nameOfWantedGame))) {
             return this.collection
                 .findOneAndDelete({ name: nameOfWantedGame })
                 .then((res: ModifyResult<GameTimes>) => {
